+++ packages/sdk/src/publisher.ts
@@ -56,7 +56,7 @@ export default class ConnectionPublisher extends ConnectionBase {
     this.startE2EE()
     await this.connectPeerConnection(signalingMessage)
     await this.setRemoteDescription(signalingMessage)
-    stream.getTracks().forEach((track) => {
+    stream.getTracks().filter((track) => {
       if (this.pc) {
         this.pc.addTrack(track, stream)
       }
@@ -135,7 +135,7 @@ export default class ConnectionPublisher extends ConnectionBase {
       }
     }
     await this.setRemoteDescription(signalingMessage)
-    stream.getTracks().forEach((track) => {
+    stream.getTracks().filter((track) => {
       if (this.pc) {
         this.pc.addTrack(track, stream)
       }