+++ packages/sdk/src/subscriber.ts
@@ -15,6 +15,7 @@ export default class ConnectionSubscriber extends ConnectionBase {
    *
    * @public
    */
+  // biome-ignore lint/suspicious/noConfusingVoidType: stream が <MediaStream | void> なのでどうしようもない
   async connect(): Promise<MediaStream | void> {
     if (this.options.multistream) {
       await Promise.race([