+++ packages/sdk/src/utils.ts
@@ -1,19 +1,19 @@
 import { unzlibSync } from 'fflate'
 
 import {
-  ConnectionOptions,
   Browser,
-  JSONType,
+  ConnectionOptions,
   DataChannelConfiguration,
   DataChannelEvent,
   DataChannelMessageEvent,
+  JSONType,
   PreKeyBundle,
-  SignalingConnectMessage,
   SignalingConnectDataChannel,
+  SignalingConnectMessage,
   SignalingEvent,
-  SignalingNotifyMetadata,
   SignalingNotifyConnectionCreated,
   SignalingNotifyConnectionDestroyed,
+  SignalingNotifyMetadata,
   TimelineEvent,
   TimelineEventLogType,
   TransportType,
@@ -223,7 +223,7 @@ export function createSignalingMessage(
     'videoAV1Params',
   ]
   const copyOptions = Object.assign({}, options)
-  ;(Object.keys(copyOptions) as (keyof ConnectionOptions)[]).forEach((key) => {
+  ;(Object.keys(copyOptions) as (keyof ConnectionOptions)[]).filter((key) => {
     if (key === 'audio' && typeof copyOptions[key] === 'boolean') {
       return
     }
@@ -254,17 +254,17 @@ export function createSignalingMessage(
   if (message.audio && hasAudioProperty) {
     message.audio = {}
     if ('audioCodecType' in copyOptions) {
-      message.audio['codec_type'] = copyOptions.audioCodecType
+      message.audio.codec_type = copyOptions.audioCodecType
     }
     if ('audioBitRate' in copyOptions) {
-      message.audio['bit_rate'] = copyOptions.audioBitRate
+      message.audio.bit_rate = copyOptions.audioBitRate
     }
   }
   const hasAudioOpusParamsProperty = Object.keys(copyOptions).some((key) => {
     return 0 <= audioOpusParamsPropertyKeys.indexOf(key)
   })
   if (message.audio && hasAudioOpusParamsProperty) {
-    if (typeof message.audio != 'object') {
+    if (typeof message.audio !== 'object') {
       message.audio = {}
     }
     message.audio.opus_params = {}
@@ -294,8 +294,8 @@ export function createSignalingMessage(
     }
   }
 
-  if (message.audio && options.audioCodecType == 'LYRA') {
-    if (typeof message.audio != 'object') {
+  if (message.audio && options.audioCodecType === 'LYRA') {
+    if (typeof message.audio !== 'object') {
       message.audio = {}
     }
 
@@ -317,19 +317,19 @@ export function createSignalingMessage(
   if (message.video && hasVideoProperty) {
     message.video = {}
     if ('videoCodecType' in copyOptions) {
-      message.video['codec_type'] = copyOptions.videoCodecType
+      message.video.codec_type = copyOptions.videoCodecType
     }
     if ('videoBitRate' in copyOptions) {
-      message.video['bit_rate'] = copyOptions.videoBitRate
+      message.video.bit_rate = copyOptions.videoBitRate
     }
     if ('videoVP9Params' in copyOptions) {
-      message.video['vp9_params'] = copyOptions.videoVP9Params
+      message.video.vp9_params = copyOptions.videoVP9Params
     }
     if ('videoH264Params' in copyOptions) {
-      message.video['h264_params'] = copyOptions.videoH264Params
+      message.video.h264_params = copyOptions.videoH264Params
     }
     if ('videoAV1Params' in copyOptions) {
-      message.video['av1_params'] = copyOptions.videoAV1Params
+      message.video.av1_params = copyOptions.videoAV1Params
     }
   }
 
@@ -353,7 +353,7 @@ export function createSignalingMessage(
       message.video = {}
     }
     if (message.video) {
-      message.video['codec_type'] = 'VP8'
+      message.video.codec_type = 'VP8'
     }
   }
 
@@ -411,7 +411,7 @@ export function trace(clientId: string | null, title: string, value: unknown): v
         // 何もしない
       }
       if (keys && Array.isArray(keys)) {
-        keys.forEach((key) => {
+        keys.filter((key) => {
           console.group(key)
           dump((record as Record<string, unknown>)[key])
           console.groupEnd()
@@ -425,18 +425,18 @@ export function trace(clientId: string | null, title: string, value: unknown): v
   }
   let prefix = ''
   if (window.performance) {
-    prefix = '[' + (window.performance.now() / 1000).toFixed(3) + ']'
+    prefix = `[${(window.performance.now() / 1000).toFixed(3)}]`
   }
   if (clientId) {
-    prefix = prefix + '[' + clientId + ']'
+    prefix = `${prefix}[${clientId}]`
   }
 
   if (console.info !== undefined && console.group !== undefined) {
-    console.group(prefix + ' ' + title)
+    console.group(`${prefix} ${title}`)
     dump(value)
     console.groupEnd()
   } else {
-    console.log(prefix + ' ' + title + '\n', value)
+    console.log(`${prefix} ${title}\n`, value)
   }
 }