+++ CHANGES.md
@@ -11,7 +11,7 @@
 
 ## develop
 
-- [FIX] 余計な依存が増えるのを防ぐために、ユーザが直接使わない型には @internalを指定して .d.ts に含まれないようにする
+- [FIX] ユーザが直接使わない型には @internal を指定して .d.ts に含まれないようにする
     - sora-js-sdk 内部で使われている E2EE と Lyra 関連の型が生成される .d.ts ファイルに含まれないようにする
     - これによって sora-js-sdk の利用者は tsconfig.json に `skipLibCheck: true` を指定せずともコンパイルができるようになる
     - @sile