+++ CHANGES.md
@@ -11,6 +11,9 @@
 
 ## develop
 
+- [CHANGE] examples を test に変更する
+  - @voluntas
+
 ## 2023.2.0
 
 **2023-12-08**